encoder/decoder: default to outputing video at internal bit depth
At daegu, "It was agreed that, when IBDI is turned on, the output of the
decoding process is, in principle, extended bit depth video".

An encoder/decoder option will be added to restore the old behaviour of
outputing Y'CbCr at the "non-IBDI" depth.

This removes TAppDecTop::xDeScalePic() and TEncGOP::xDeScalePic()

NB, it is now important that the output file in the decoder is not
opened before the SPS has been parsed to determine the internal
bitdepth.  (It needs this information to know the output filetype)

From: David Flynn <davidf@rd.bbc.co.uk>