build/linux,msvc: do not define NDEBUG in any builds
NDEBUG is a preprocessor macro that (amongst other things) causes
assert(0) to become a no-op.  The premise being that assert(x) offers
such a strong guarantee that x is true, there is no need to test for
it in release builds.  Unfortunately, in the this software, the
guarantee isn't always very strong, with assert sometimes being used
to replace formal error handling.

Furthermore, due to the slow speed of the software at the best of times,
it doesn't make sense to have build configurations along the lines of
 either (debug): -O0 -g
 or   (release): -O3 -NDEBUG
Rather, what is actually desired is "debugging with and without
optimizations".

From: David Flynn <davidf@rd.bbc.co.uk>